backend/aquaguard-svc/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.7.0
pydantic-settings==2.3.0
paho-mqtt==2.1.0
structlog==24.2.0
